Create make shopping
This application add to sqlite databases the differents Items of the shop for professionnal and  allow to add items in the basket of the clients who desire to realize a chopping list. Once this shopping list is finalized a qrcode will appear with the total of the shopping list and the list itself for each ordering.

NB: I precise that I did not finish for now this application it is just my copy draft.